requirements.txt
astroid==2.5.1
bidict==0.21.2
certifi==2020.12.5
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.7.2
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yapf==0.31.0
